+++ test/client.ts
@@ -45,7 +45,7 @@ import {
 
 import {
   QueryManager,
-} from '../src/QueryManager';
+} from '../src/core/QueryManager';
 
 import {
   createNetworkInterface,
